requirements.txt
discord.py
openai>=1.0.0
pymongo
python-dotenv
rcon